+++ Web/src/main/ui/material/src/app/components/manage/account-edit/account-edit.component.ts
@@ -1,10 +1,11 @@
-import { Component, OnInit } from '@angular/core';
-import {Routes, RouterModule, Router, ActivatedRoute} from "@angular/router";
-import { AccountService } from '../../../services/account.service';
-import { OrgService } from '../../../services/org.service';
-import { Account } from '../../../models/account.model';
-import { Handler } from '../../dialogs/handler/handler';
-
+import{Component, OnInit, Inject}from '@angular/core';
+import {Routes, RouterModule, Router, ActivatedRoute}from "@angular/router";
+import {AccountService}from '../../../services/account.service';
+import {OrgService}from '../../../services/org.service';
+import { Account}from '../../../models/account.model';
+import {Handler}from '../../dialogs/handler/handler';
+import {VERSION, MatDialog, MatDialogRef, MatSnackBar }from '@angular/material';
+import {DeleteDialogComponent}from '../../dialogs/delete-dialog/delete-dialog.component';
 
 @Component({
   selector: 'app-account-edit',
@@ -21,7 +22,7 @@ export class AccountEditComponent implements OnInit {
   cloudTypes = ['AWS','DIGITAL_OCEAN','GCP','AZURE','PRIVATE_CLOUD','VMWARE','OPENSTACK','OTHER'];
   accountTypes = ['Git', 'GitHub', 'Jira', 'BitBucket', 'GitLab', 'Microsoft_TFS_Git', 'Microsoft_VSTS_Git', 'Local', 'AWS', 'Slack' , 'Email' , 'Self_Hosted'];
    AWSREGIONS = ['us-east-1','us-east-2','us-west-1','us-west-2','ca-central-1','eu-central-1','eu-west-1','eu-west-2','eu-west-3','ap-northeast-1','ap-northeast-2','ap-northeast-3','ap-southeast-1','ap-southeast-2','ap-southeast-1','sa-east-1'];
-  constructor(private accountService: AccountService, private orgService: OrgService, private route: ActivatedRoute, private router: Router, private handler: Handler) { }
+  constructor(private accountService: AccountService, private orgService: OrgService, private route: ActivatedRoute, private router: Router, private handler: Handler, public dialog: MatDialog) { }
 
   ngOnInit() {
     this.route.params.subscribe(params => {
@@ -60,20 +61,30 @@ export class AccountEditComponent implements OnInit {
     });
   }
 
-  delete() {
-    this.handler.activateLoader();
-    this.accountService.delete(this.entry).subscribe(results => {
-      this.handler.hideLoader();
-      if (this.handler.handle(results)) {
-        return;
-      }
-      this.router.navigate(['/app/accounts']);
-    }, error => {
-      this.handler.hideLoader();
-      this.handler.error(error);
+ delete() {
+    let dialogRef = this.dialog.open(DeleteDialogComponent, {
+        data: {
+            entry: this.entry
+        }
     });
-  }
 
+    dialogRef.afterClosed().subscribe(result => {
+        if (result != null) {
+            this.handler.activateLoader();
+            this.accountService.delete(this.entry).subscribe(results => {
+                this.handler.hideLoader();
+                if (this.handler.handle(results)) {
+                    return;
+                }
+                this.router.navigate(['/app/accounts']);
+            }, error => {
+                this.handler.hideLoader();
+                this.handler.error(error);
+
+            });
+        }
+    });
+}
   getOrgs() {
     this.handler.activateLoader();
     this.orgService.getByUser().subscribe(results => {
