+++ Web/src/main/ui/material/src/app/components/projects/projects-edit/projects-edit.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Inject } from '@angular/core';
 import {Routes, RouterModule, Router, ActivatedRoute} from "@angular/router";
 import { ProjectService } from '../../../services/project.service';
 import { OrgService } from '../../../services/org.service';
@@ -6,6 +6,8 @@ import { AccountService } from '../../../services/account.service';
 import { Project } from '../../../models/project.model';
 import { Handler } from '../../dialogs/handler/handler';
 import { APPCONFIG } from '../../../config';
+import {VERSION, MatDialog, MatDialogRef, MatSnackBar }from '@angular/material';
+import {DeleteDialogComponent}from '../../dialogs/delete-dialog/delete-dialog.component';
 
 @Component({
   selector: 'app-projects-edit',
@@ -20,7 +22,7 @@ export class ProjectsEditComponent implements OnInit {
   project: Project = new Project();
   accounts;
   public AppConfig: any;
-  constructor(private projectService: ProjectService, private accountService: AccountService, private orgService: OrgService, private route: ActivatedRoute, private router: Router, private handler: Handler) { }
+  constructor(private projectService: ProjectService, private accountService: AccountService, private orgService: OrgService, private route: ActivatedRoute, private router: Router, private handler: Handler, , public dialog: MatDialog) { }
 
   ngOnInit() {
     this.AppConfig = APPCONFIG;
@@ -61,20 +63,30 @@ export class ProjectsEditComponent implements OnInit {
       this.handler.error(error);
     });
   }
+ delete() {
+     let dialogRef = this.dialog.open(DeleteDialogComponent, {
+         data: {
+             project: this.project
+         }
+     });
 
-  delete() {
-    console.log(this.project);
-    this.projectService.delete(this.project).subscribe(results => {
-      this.handler.hideLoader();
-      if (this.handler.handle(results)) {
-        return;
-      }
-      this.router.navigate(['/app/projects']);
-    }, error => {
-      this.handler.hideLoader();
-      this.handler.error(error);
-    });
-  }
+     dialogRef.afterClosed().subscribe(result => {
+         if (result != null) {
+             this.handler.activateLoader();
+             this.projectService.delete(this.project).subscribe(results => {
+                 this.handler.hideLoader();
+                 if (this.handler.handle(results)) {
+                     return;
+                 }
+                 this.router.navigate(['/app/projects']);
+             }, error => {
+                 this.handler.hideLoader();
+                 this.handler.error(error);
+
+             });
+         }
+     });
+ }
 
   getOrgs() {
     this.orgService.getByUser().subscribe(results => {
